+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [Unreleased]
 
 ### Added
-- 🆒 1-value always-on focus screen (with brightness adjust for Android)
+- 🆒 1-value always-on focus screen with font color choice (and brightness adjust for Android)
 
 ### Changed
 - much improved timezone keyword search